+++ .eslintignore
@@ -1,3 +1,3 @@
-# Ignore built files except build/index.js
+# Ignore built files except build/index.jsx
 build/*
-!build/index.js
+!build/index.jsx
