+++ src/App.jsx
@@ -1,9 +1,9 @@
-import React from "react"
-import styled from "styled-components"
+import React from 'react';
+import styled from 'styled-components';
 
-import { GlobalStyles } from "./components/utils/theme"
+import { GlobalStyles } from './components/utils/theme';
 
-import LandingPage from "./components/pages/LandingPage"
+import LandingPage from './components/pages/LandingPage';
 
 const LandingPageContainer = styled.div`
   position: absolute;
@@ -11,17 +11,17 @@ const LandingPageContainer = styled.div`
   left: 0;
   width: 100%;
   height: 100%;
-`
+`;
 
 function App() {
   return (
-    <React.Fragment>
+    <>
       <GlobalStyles />
       <LandingPageContainer>
         <LandingPage />
       </LandingPageContainer>
-    </React.Fragment>
-  )
+    </>
+  );
 }
 
-export default App
+export default App;
