+++ src/components/hooks/index.js
@@ -1,2 +1,2 @@
-export { default as useToggleVisibility } from "./useToggleVisibility"
-export { default as useInput } from "./useInput"
+export { default as useToggleVisibility } from './useToggleVisibility';
+export { default as useInput } from './useInput';
