+++ src/components/hooks/useToggleVisibility.js
@@ -1,23 +1,22 @@
-import { useState, useEffect, useRef } from "react"
+import { useState, useEffect, useRef } from 'react';
 
 const useToggleVisibility = (initialState = false, timeoutTime) => {
-  const [visible, setVisibility] = useState(initialState)
-  const visibleRef = useRef(visible)
-  visibleRef.current = visible
+  const [visible, setVisibility] = useState(initialState);
+  const visibleRef = useRef(visible);
+  visibleRef.current = visible;
 
   const toggleVisibility = () => {
-    setVisibility(!visible)
-  }
+    setVisibility(!visible);
+  };
 
   useEffect(() => {
-    const timer =
-      timeoutTime && visibleRef.current
-        ? setTimeout(() => toggleVisibility(), timeoutTime)
-        : null
-    return () => clearTimeout(timer)
-  })
+    const timer = timeoutTime && visibleRef.current
+      ? setTimeout(() => toggleVisibility(), timeoutTime)
+      : null;
+    return () => clearTimeout(timer);
+  });
 
-  return [visible, toggleVisibility, setVisibility]
-}
+  return [visible, toggleVisibility, setVisibility];
+};
 
-export default useToggleVisibility
+export default useToggleVisibility;
