--- src/components/utils/expressInteraction.js
@@ -1,3 +1,0 @@
-export const URL_STUBS = {
-  ADD_VISITOR: "/visitors/add"
-}
