+++ src/components/utils/helperFunctions.js
@@ -1,25 +1,25 @@
-const formatToUnit = (element, defaultUnit = "px") => {
-  const stringElement = String(element)
+const formatToUnit = (element, defaultUnit = 'px') => {
+  const stringElement = String(element);
   const values = [
-    "cm",
-    "mm",
-    "Q",
-    "in",
-    "pc",
-    "pt",
-    "px",
-    "em",
-    "ex",
-    "rem",
-    "lh",
-    "vw",
-    "vh",
-    "vmin",
-    "vmax"
-  ]
-  return values.some(value => stringElement.includes(value))
+    'cm',
+    'mm',
+    'Q',
+    'in',
+    'pc',
+    'pt',
+    'px',
+    'em',
+    'ex',
+    'rem',
+    'lh',
+    'vw',
+    'vh',
+    'vmin',
+    'vmax',
+  ];
+  return values.some((value) => stringElement.includes(value))
     ? element
-    : `${element}${defaultUnit}`
-}
+    : `${element}${defaultUnit}`;
+};
 
-export { formatToUnit }
+export { formatToUnit };
